Move location of monkey patch
